requirements.txt
cuquantum
cuquantum-python
matplotlib
numpy
qiskit-aer-gpu
ipykernel
pylatexenc
matplotlib
qiskit[visualization]
qiskit-qasm3-import